requirements.txt
dataclasses==0.6
Django==2.1.2
djangorestframework==3.9.0
djongo==1.2.30
pymongo==3.7.2
pytz==2018.6
sqlparse==0.2.4
